module 3. task 6

+++ src/main/java/lang/print/gaps/task6/FloatTypeCasting.java
@@ -2,7 +2,7 @@ package lang.print.gaps.task6;
 
 public class FloatTypeCasting {
     public void roundNumber(float numberToBeRounded) {
-        double r = Math.round(numberToBeRounded);
+        int r = Math.round(numberToBeRounded);
         System.out.println(r);
 
 //    }
